后台授权管理

+++ project-cluster-service/coin-exchange/coin-admin/admin-service/src/main/java/io/jnyou/domain/SysRolePrivilege.java
@@ -25,9 +25,9 @@ public class SysRolePrivilege {
 
     @TableField(value = "role_id")
     @ApiModelProperty(value = "")
-    private Integer roleId;
+    private Long roleId;
 
     @TableField(value = "privilege_id")
     @ApiModelProperty(value = "")
-    private Integer privilegeId;
+    private Long privilegeId;
 }
